requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
email-validator>=2.2.0